# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Maratona de Resultados API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List events
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: quarter
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/close:
    post:
      operationId: closeEvent
      tags: [events]
      summary: Close event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForceCloseInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/reopen:
    post:
      operationId: reopenEvent
      tags: [events]
      summary: Reopen event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{id}/participants:
    get:
      operationId: getEventParticipants
      tags: [events]
      summary: List event participants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventParticipant"
    post:
      operationId: addEventParticipant
      tags: [events]
      summary: Add participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventParticipantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventParticipant"
  /events/{id}/participants/{participantId}:
    delete:
      operationId: removeEventParticipant
      tags: [events]
      summary: Remove participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/criteria:
    get:
      operationId: getEventCriteria
      tags: [events]
      summary: Get event criteria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventCriterion"
    put:
      operationId: updateEventCriteria
      tags: [events]
      summary: Update event active criteria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCriteriaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventCriterion"
...
  /exports/event-results:
    get:
      operationId: exportEventResults
      tags: [exports]
      summary: Export event results CSV
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExport"
  /exports/quarterly-results:
    get:
      operationId: exportQuarterlyResults
      tags: [exports]
      summary: Export quarterly results CSV
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: quarter
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExport"
  /exports/ranking:
    get:
      operationId: exportRanking
      tags: [exports]
      summary: Export ranking CSV
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: quarter
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExport"
  /exports/caju-bonuses:
    get:
      operationId: exportCajuBonuses
      tags: [exports]
      summary: Export Caju bonus report CSV
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: quarter
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExport"
components:
  schemas:
...
    Event:
      type: object
      required: [id, name, startDate, endDate, year, quarter, status]
      properties:
        id:
          type: integer
        externalId:
          type: ["string", "null"]
        name:
          type: string
        clientName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        year:
          type: integer
        quarter:
          type: integer
        status:
          type: string
        eventWeight:
          type: number
        forcedClosed:
          type: boolean
        participantCount:
          type: integer
        evaluationProgress:
          type: number
        averageScore:
          type: ["number", "null"]
        createdAt:
          type: string
    EventDetail:
      type: object
      required: [id, name, startDate, endDate, year, quarter, status]
      properties:
        id:
          type: integer
        name:
          type: string
        clientName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        year:
          type: integer
        quarter:
          type: integer
        status:
          type: string
        forcedClosed:
          type: boolean
        forcedCloseReason:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/EventParticipant"
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/EventCriterion"
        evaluationMatrix:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationMatrixRow"
        results:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeEventResult"
    EventInput:
      type: object
      required: [name, startDate, endDate, year, quarter]
      properties:
        name:
          type: string
        clientName:
          type: string
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        year:
          type: integer
        quarter:
          type: integer
    EventUpdate:
      type: object
      properties:
        name:
          type: string
        clientName:
          type: string
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
    ForceCloseInput:
      type: object
      properties:
        forced:
          type: boolean
        reason:
          type: string
    EventParticipant:
      type: object
      required: [id, eventId, employeeId, employeeName, functionName]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        functionName:
          type: string
        teamName:
          type: ["string", "null"]
        confirmed:
          type: boolean
    EventParticipantInput:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer
        functionName:
          type: string
        teamName:
          type: string
...
    EventCriterion:
      type: object
      required: [id, eventId, criterionId, criterionName, active, normalizedWeight]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        criterionId:
          type: integer
        criterionName:
          type: string
        responsibleAreaName:
          type: ["string", "null"]
        active:
          type: boolean
        originalWeight:
          type: number
        weightOverride:
          type: ["number", "null"]
        normalizedWeight:
          type: number
    EventCriteriaUpdate:
      type: object
      required: [activeCriterionIds]
      properties:
        activeCriterionIds:
          type: array
          items:
            type: integer
...
    EvaluationMatrixRow:
      type: object
      required: [employeeId, employeeName, criteria]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationMatrixCell"
    EvaluationMatrixCell:
      type: object
      required: [criterionId, criterionName, status]
      properties:
        criterionId:
          type: integer
        criterionName:
          type: string
        status:
          type: string
        averageScore:
          type: ["number", "null"]
        calibratedScore:
          type: ["number", "null"]
...
    EmployeeEventResult:
      type: object
      required: [employeeId, employeeName, eventId, eventScore]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        eventId:
          type: integer
        eventScore:
          type: number
        projectedPlatoon:
          type: ["string", "null"]
        criteriaDetails:
          type: array
          items:
            $ref: "#/components/schemas/CriterionScoreDetail"
    CriterionScoreDetail:
      type: object
      required: [criterionId, criterionName, normalizedWeight]
      properties:
        criterionId:
          type: integer
        criterionName:
          type: string
        averageScore:
          type: ["number", "null"]
        calibratedScore:
          type: ["number", "null"]
        scoreUsed:
          type: ["number", "null"]
        scorePercentual:
          type: ["number", "null"]
        normalizedWeight:
          type: number
        weightedContribution:
          type: ["number", "null"]
...
    CsvExport:
      type: object
      required: [filename, data]
      properties:
        filename:
          type: string
        data:
          type: string